requirements.txt
Django
djangorestframework
django_apscheduler
requests
bs4